Folder reorganized, dockerfile & requirements file added
