+++ app/etl_project/.env
@@ -1,0 +1,13 @@
+api_key=c4b96f23f578ecde02588a9f5e7478a3
+
+SERVER_NAME=etl-nba.caly07jx47gb.us-east-1.rds.amazonaws.com
+DATABASE_NAME=nba
+DB_USERNAME=postgres
+DB_PASSWORD='u4>haVVrLi3?H{zWV8<8FnX#F)&9'
+PORT=5432
+
+LOGGING_SERVER_NAME=etl-nba.caly07jx47gb.us-east-1.rds.amazonaws.com
+LOGGING_DATABASE_NAME=nba
+LOGGING_USERNAME=postgres
+LOGGING_PASSWORD='u4>haVVrLi3?H{zWV8<8FnX#F)&9'
+LOGGING_PORT=5432
