+++ app/etl_project/logs/nba_1693968921.3978388.log
@@ -1,0 +1,8 @@
+2023-09-05 22:55:21,673 - nba - INFO - Starting pipeline run
+2023-09-05 22:55:21,680 - nba - INFO - Getting pipeline environment variables
+2023-09-05 22:55:21,680 - nba - INFO - Creating NBA API client
+2023-09-05 22:55:21,681 - nba - INFO - Extracting data from NBA API client
+2023-09-05 22:55:21,681 - nba - INFO - Extracting games data from NBA API client
+2023-09-05 22:55:23,754 - nba - INFO - Extracting players data from NBA API client
+2023-09-05 22:56:28,303 - nba - INFO - Extracting players statistics data from NBA API client
+2023-09-05 22:56:29,202 - nba - INFO - Pipeline run failed. Error message: Failed to extract data from NBA API. Status Code: 200. Response: {"get":"teams","parameters":{"league":"standard"},"errors":{"rateLimit":"Too many requests. Your rate limit is 10 requests per minute."},"results":0,"paging":{"current":1,"total":1},"response":[]}
