+++ app/etl_project/logs/nba_1693970310.1956573.log
@@ -1,0 +1,3 @@
+2023-09-05 23:18:30,197 - nba - INFO - Getting pipeline environment variables
+2023-09-05 23:18:30,198 - nba - INFO - Creating NBA API client
+2023-09-05 23:18:30,199 - nba - INFO - Pipeline run failed. Error message: name 'postgresql_logging_client' is not defined
