+++ app/etl_project/logs/nba_1694004673.45831.log
@@ -1,0 +1,15 @@
+2023-09-06 08:51:13,770 - nba - INFO - Starting pipeline run
+2023-09-06 08:51:13,770 - nba - INFO - Getting pipeline environment variables
+2023-09-06 08:51:13,770 - nba - INFO - Creating NBA API client
+2023-09-06 08:51:13,770 - nba - INFO - Extracting data from NBA API client
+2023-09-06 08:51:13,770 - nba - INFO - Extracting games data from NBA API client
+2023-09-06 08:51:15,283 - nba - INFO - Extracting players data from NBA API client
+2023-09-06 08:52:31,824 - nba - INFO - Extracting players statistics data from NBA API client
+2023-09-06 08:53:38,575 - nba - INFO - Extracting standings data from NBA API client
+2023-09-06 08:53:39,440 - nba - INFO - Transforming games df
+2023-09-06 08:53:39,534 - nba - INFO - Transforming players statistics
+2023-09-06 08:53:39,721 - nba - INFO - Transforming standings
+2023-09-06 08:53:39,721 - nba - INFO - Loading data to postgres
+2023-09-06 08:53:39,721 - nba - INFO - Loading games data to postgres
+2023-09-06 08:53:40,252 - nba - INFO - Loading standings data to postgres
+2023-09-06 08:53:40,301 - nba - INFO - Loading players statistics data to postgres
