+++ app/etl_project/pipelines/nba.yaml
@@ -1,0 +1,5 @@
+name: nba
+config: 
+  season: 2022
+  league: "standard"
+  log_folder_path: "./etl_project/logs"
