--- etl_project/logs/nba_1693875567.369715.log
@@ -1,14 +1,0 @@
-2023-09-04 20:59:27,487 - nba - INFO - Starting pipeline run
-2023-09-04 20:59:27,497 - nba - INFO - Getting pipeline environment variables
-2023-09-04 20:59:27,498 - nba - INFO - Creating NBA API client
-2023-09-04 20:59:27,498 - nba - INFO - Extracting data from NBA API client
-2023-09-04 20:59:27,499 - nba - INFO - Extracting games data from NBA API client
-2023-09-04 20:59:28,795 - nba - INFO - Extracting players data from NBA API client
-2023-09-04 21:00:15,256 - nba - INFO - Extracting players statistics data from NBA API client
-2023-09-04 21:01:09,796 - nba - INFO - Extracting standings data from NBA API client
-2023-09-04 21:01:10,514 - nba - INFO - Transforming games df
-2023-09-04 21:01:10,541 - nba - INFO - Transforming players statistics
-2023-09-04 21:01:10,625 - nba - INFO - Transforming standings
-2023-09-04 21:01:10,626 - nba - INFO - Loading data to postgres
-2023-09-04 21:01:10,626 - nba - INFO - Loading games data to postgres
-2023-09-04 21:01:10,925 - nba - INFO - Loading standings data to postgres
